# docs/integration/quotation-draft-v1.1.openapi.yaml
openapi: 3.0.3
info:
  title: Genesis_CRM Integration API — Quotation Draft Extension
  version: 1.1.0
  description: |
    Integration API v1 的向后兼容报价能力扩展，能力标识为 `quotation-draft.v1`。
    基础契约 `integration-v1.openapi.yaml`、`contractVersion: "1.0"` 与其冻结哈希保持不变。

    AutoForceAI 只生成可编辑建议；用户明确确认后，本扩展才允许在 Genesis_CRM
    创建正式的 `draft` 报价。Genesis 是报价编号、状态、行金额、总额和 PDF 的权威来源。

    所有端点必须使用服务凭证和显式 `X-Project-Id`。写端点还必须携带
    `Idempotency-Key`。不得使用用户 JWT、邮箱或客户姓名替代外部客户引用。
  contact:
    name: AutoForceAI / Genesis_CRM Integration Team
x-base-contract-version: "1.0"
x-capability: quotation-draft.v1

servers:
  - url: http://localhost:5000/api/integrations/v1
    description: 本地开发

tags:
  - name: quotations

paths:
  /customers/{externalRef}/quotation-drafts:
    post:
      tags: [quotations]
      summary: 为已关联客户幂等创建报价草稿
      operationId: createQuotationDraftByExternalRef
      description: |
        只创建 `draft`。Genesis 按 externalRef 定位当前项目客户，忽略任何客户端金额汇总，
        根据 `quantity × unitPrice` 重算行金额与总额。相同幂等键和相同载荷返回首次结果；
        相同键但载荷不同返回 `409 CONFLICT`。
      security:
        - serviceToken: ["quotations:draft"]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ExternalRef'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuotationDraftRequest'
            example:
              schemaVersion: "1.0"
              sourceSystem: autoforce
              title: "5000 pcs Recycled Canvas Tote Bags"
              currency: USD
              items:
                - productName: "Recycled Canvas Tote Bag"
                  model: "GB-2026-A"
                  quantity: 5000
                  unitPrice: 1.2
              validityDate: "2026-10-24T00:00:00.000Z"
              paymentTerms: "30% deposit, 70% before shipment"
              leadTime: "45 days after sample approval"
              moq: "1000 pcs"
              notes: "FOB Xiamen; final color subject to approved sample."
              markCustomerAsQuoting: true
              proposalTrace:
                proposalId: "quote-proposal:9d4dd8b0"
                generatedBy: autoforce_ai
                model: "configured-model"
                sources:
                  - kind: lead
                    referenceId: "lead:1024"
                    title: "Website inquiry"
                  - kind: knowledge
                    referenceId: "kb-doc:88#chunk:4"
                    title: "Canvas bag MOQ and lead-time guide"
      responses:
        '201':
          description: 报价草稿已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationResponse'
        '200':
          description: 相同幂等键的成功重放，返回首次创建的同一报价
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'

  /quotations/{quotationId}:
    get:
      tags: [quotations]
      summary: 读取项目内报价权威详情
      operationId: getIntegrationQuotation
      security:
        - serviceToken: ["quotations:read"]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/QuotationId'
      responses:
        '200':
          description: 报价详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'
        '404':
          $ref: '#/components/responses/NotFound'

  /quotations/{quotationId}/pdf:
    get:
      tags: [quotations]
      summary: 下载由 Genesis 服务端生成的报价 PDF
      operationId: downloadIntegrationQuotationPdf
      description: |
        PDF 只从数据库中的权威报价生成。相同 `quotationId + version` 必须返回稳定内容和 ETag；
        报价更新后 version 增加，旧 ETag 失效。draft 报价必须带 DRAFT 水印。
      security:
        - serviceToken: ["quotations:read"]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/QuotationId'
        - name: If-None-Match
          in: header
          required: false
          schema: { type: string }
          description: 上一次响应的 ETag。
      responses:
        '200':
          description: 报价 PDF
          headers:
            ETag:
              schema: { type: string }
              description: 基于 quotationId、version 与 PDF 内容的稳定强校验标识。
            Content-Disposition:
              schema: { type: string }
              description: 安全 ASCII fallback 与 RFC 5987 UTF-8 文件名。
            X-Quotation-Version:
              schema: { type: integer, minimum: 1 }
            X-Content-Type-Options:
              schema: { type: string, enum: [nosniff] }
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '304':
          description: If-None-Match 命中，报价版本未变化
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/PdfUnavailable'

components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Genesis 管理员签发并绑定 project/scope 的服务凭证。

  parameters:
    ProjectId:
      name: X-Project-Id
      in: header
      required: true
      schema:
        type: string
        pattern: '^[a-f\d]{24}$'
      description: 必须显式携带且在服务凭证绑定的项目集合内。
    ExternalRef:
      name: externalRef
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
      description: "已通过 Integration API 关联的稳定外部客户引用，如 `lead:1024`。"
    QuotationId:
      name: quotationId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-f\d]{24}$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: 网络重试必须复用原键；用户修改建议后重新确认必须生成新键。

  schemas:
    CreateQuotationDraftRequest:
      type: object
      additionalProperties: false
      required: [schemaVersion, sourceSystem, title, items, currency]
      properties:
        schemaVersion:
          type: string
          enum: ["1.0"]
        sourceSystem:
          type: string
          enum: [autoforce]
        title:
          type: string
          minLength: 1
          maxLength: 200
        items:
          type: array
          minItems: 1
          maxItems: 200
          items:
            $ref: '#/components/schemas/CreateQuotationItem'
        currency:
          $ref: '#/components/schemas/QuotationCurrency'
        validityDate:
          type: string
          format: date-time
          nullable: true
        paymentTerms:
          type: string
          maxLength: 300
        leadTime:
          type: string
          maxLength: 200
        moq:
          type: string
          maxLength: 120
        notes:
          type: string
          maxLength: 5000
        markCustomerAsQuoting:
          type: boolean
          default: false
        proposalTrace:
          $ref: '#/components/schemas/ProposalTrace'

    CreateQuotationItem:
      type: object
      additionalProperties: false
      required: [productName, quantity, unitPrice]
      properties:
        productName:
          type: string
          minLength: 1
          maxLength: 200
        model:
          type: string
          maxLength: 200
        quantity:
          type: number
          minimum: 0
          exclusiveMinimum: true
          maximum: 1000000000
        unitPrice:
          type: number
          minimum: 0
          maximum: 1000000000
      description: "`amount` 不属于输入；出现未知字段必须拒绝或剥离后不参与计算。"

    ProposalTrace:
      type: object
      additionalProperties: false
      required: [proposalId, generatedBy, sources]
      properties:
        proposalId:
          type: string
          minLength: 1
          maxLength: 128
        generatedBy:
          type: string
          enum: [autoforce_ai]
        model:
          type: string
          maxLength: 120
          description: 脱敏模型标识，不含密钥、供应商请求 ID 或提示词。
        sources:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/SourceReference'

    SourceReference:
      type: object
      additionalProperties: false
      required: [kind, referenceId]
      properties:
        kind:
          type: string
          enum: [lead, knowledge, customer, quotation]
        referenceId:
          type: string
          minLength: 1
          maxLength: 200
        title:
          type: string
          maxLength: 300

    QuotationCurrency:
      type: string
      enum: [USD, EUR, GBP, CNY, JPY, HKD, AUD, CAD, CHF, SGD, AED, NZD]

    QuotationStatus:
      type: string
      enum: [draft, sent, negotiating, accepted, rejected, expired]

    QuotationItem:
      type: object
      additionalProperties: false
      required: [productName, quantity, unitPrice, amount]
      properties:
        productName: { type: string }
        model: { type: string, nullable: true }
        quantity: { type: number }
        unitPrice: { type: number }
        amount:
          type: number
          description: Genesis 计算的 `round(quantity × unitPrice, 2)`。

    Quotation:
      type: object
      additionalProperties: false
      required:
        - quotationId
        - quotationNo
        - customerId
        - externalRef
        - title
        - items
        - currency
        - totalAmount
        - status
        - version
        - createdAt
        - updatedAt
      properties:
        quotationId: { type: string }
        quotationNo: { type: string }
        customerId: { type: string }
        externalRef: { type: string }
        title: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuotationItem'
        currency:
          $ref: '#/components/schemas/QuotationCurrency'
        totalAmount:
          type: number
          description: Genesis 计算的行金额之和。
        validityDate: { type: string, format: date-time, nullable: true }
        paymentTerms: { type: string, nullable: true }
        leadTime: { type: string, nullable: true }
        moq: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        status:
          $ref: '#/components/schemas/QuotationStatus'
        version:
          type: integer
          minimum: 1
        proposalTrace:
          allOf:
            - $ref: '#/components/schemas/ProposalTrace'
          nullable: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    QuotationResponse:
      type: object
      additionalProperties: false
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/Quotation'

    Error:
      type: object
      additionalProperties: false
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items: { type: object }
            requestId: { type: string }

  responses:
    Unauthorized:
      description: token 缺失、无效、已撤销或已过期
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            success: false
            error: { code: UNAUTHORIZED, message: 服务凭证无效或已失效 }
    ForbiddenScope:
      description: 凭证缺少所需 scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            success: false
            error: { code: FORBIDDEN_SCOPE, message: "凭证缺少所需 scope: quotations:draft" }
    NotFound:
      description: 客户或报价不存在，或不属于当前项目
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            success: false
            error: { code: NOT_FOUND, message: 资源不存在 }
    ValidationError:
      description: 请求字段不合法，消费方不得自动重试
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            success: false
            error: { code: VALIDATION_ERROR, message: 报价字段校验未通过 }
    IdempotencyConflict:
      description: 相同幂等键已被不同载荷使用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            success: false
            error: { code: CONFLICT, message: Idempotency-Key 已被不同载荷使用 }
    RateLimited:
      description: 请求频率过高；响应应携带 Retry-After
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            success: false
            error: { code: RATE_LIMITED, message: 请求频率过高 }
    PdfUnavailable:
      description: PDF 暂时无法生成；报价草稿本身仍然有效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            success: false
            error: { code: PDF_UNAVAILABLE, message: PDF 暂时无法生成，请稍后重试 }